pyproject.toml
[tool.poetry]
name = "derisk-research"
version = "0.1.0"
description = ""
authors = ["Lukas Petrasek <lukas@carmine.finance>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
streamlit = "^1.35.0"
pandas = "^2.2.2"
starknet-py = "^0.22.0"
plotly = "^5.22.0"
fastparquet = "^2024.5.0"
google-cloud-storage = "^2.16.0"
psycopg2-binary = "^2.9.9"
matplotlib = { version = "3.8.0", optional = true }
tqdm = { version = "4.65.0", optional = true }
yfinance = { version = "0.2.38", optional = true }
dill = "^0.3.8"
gcsfs = "^2024.6.1"
pre-commit = "^3.8.0"
pydantic = "^2.9.2"
pytest = "^8.3.3"
python-dotenv = "^1.0.1"
alembic = "^1.13.3"
sqlalchemy-utils = "^0.41.2"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# A list of all of the optional dependencies
[tool.poetry.extras]
liquidation = ["matplotlib", "tqdm", "yfinance"]

[tool.isort]
profile = "black"

